Handle bundle output path when creating bundle target
It seems there is no good way to get the path to the chunk(s) in a
bundle when we just use the rollup API to generate chunks. So we're left
with having to deduce the path ourselves using the output options.

We force the user to specify one of 'file' or 'dir'. This has an added
bonus of preventing the default rollout behavior which could potentially
overwrite an input target.