Fix prefix resolution (again)
There are two cases to consider: one with an import context and one
without. In the case with an import context, the importer is already
resolved, so we have to qualify an absolute source ID with the prefix so
the resolution can work out. In the remaining case, we can just prepend
it after resolution since the prefix is not introduced during the path
resolution.