Improve bundle refs import dependency ordering
This change greatly improves how we order refs. It turns out that the
load order didn't really map well to the actual dependency order in some
cases. We now use the list of external imports provided by the output
chunks to map to the references generated for those imports. For source
imports that do not map to external imports, we just fall back to using
the load order.