Prevent creation of empty ref groups
An empty ref group tells the manifest plugin to make all previous refs
in the same group eligible for deletion. We only need to create a ref
group if it is non-empty. The manifest plugin will decide if a previous
group's refs should be deleted if it is not present.